pyproject.toml
[build-system]
requires = ["maturin>=1.2,<2.0"]
build-backend = "maturin"

[project]
name = "riir-airflow"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "fastapi[standard]>=0.112.0",
    "pydantic-settings>=2.6.1",
    "apache-airflow>=2.9.1,<2.10",
    "sh>=2.0.6",
    "orjson>=3.10.3",
    "nicegui>=2.8.1",
]
readme = "README.md"
requires-python = ">= 3.12"

[project.scripts]
riir-airflow = 'riir_airflow.__main__:cli_app'

[tool.maturin]
python-source = "python"
module-name = "riir_airflow._lowlevel"
features = ["pyo3/extension-module"]

[tool.uv]
dev-dependencies = [
    "mike>=2.1.3",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.5.38",
    "mkdocstrings-python>=1.11.1",
    "pre-commit>=3.8.0",
    "pytest>=8.3.3",
    "ruff>=0.6.8",
]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "B", "I"]
